Add StatsHandler for grpc-gateway
